cleaned up examples and added documentation
